refactor(nitro): move route rule segment unification into nitro-server
